+++ application/modules/editor/src/Terminology/SearchCollection.php
@@ -169,7 +169,7 @@ class SearchCollection
             ->setIntegrityCheck(false)
             ->from('terms_term')
             /* @phpstan-ignore-next-line since removed anyway in a short */
-            ->where(FEATURE_TRANSLATE_4673_ENABLE
+            ->where(defined('FEATURE_TRANSLATE_4673_ENABLE') && FEATURE_TRANSLATE_4673_ENABLE
                 ? ('term ' . $compareWith . '?')
                 : ('lower(term) ' . $compareWith . ' lower(?) COLLATE utf8mb4_bin'), $this->query)
             ->where('collectionId = ?', $this->collectionId)