+++ public/modules/editor/js/app/view/segments/new/EditorNew.js
@@ -230,6 +230,13 @@ Ext.define('Editor.view.segments.new.EditorNew', {
 
         // Set height for both editor container and editor row panel
         me.setHeight(targetHeight).up().setHeight(targetHeight);
+
+        // Make sure grid row node's height fits the editor's row height
+        // to prevent editor from overlapping next row node
+        let recNode = me.editingPlugin.view.getNodeByRecord(me.currentlyEditingRecord);
+        if (recNode) {
+            recNode.style.height = targetHeight + 'px';
+        }
     },
     getTargetHeightUsage: function() {
         return this.editor._editor.ui.view.element.clientHeight + this.statusStrip.getHeight() + 2;